refactor(lint): share CSS declaration parsing across style rules
no_dupe_style_properties.rs and no_shorthand_style_property_overrides.rs
each carried a near-identical ~150-180 line CSS declaration parser
(parse_style_decls / push_decl / extract_inline_style_decls /
extract_from_expr / extract_css_decl(s)_from_literal / _from_template),
differing only in variable names and whether they returned
(name, start, end) triples or (name, start) pairs. Extract the shared
logic into rules/shared/style_decls.rs (triples as the canonical shape)
and have both rules consume it, deriving pairs via a trivial map where
needed. Diagnostic output is unchanged.
