+++ crates/rsvelte_lint/src/rules/shared/mod.rs
@@ -1,0 +1,3 @@
+//! Logic shared across two or more rule implementations. Not itself a `Rule`.
+
+pub mod style_decls;
